requirements.txt
fastapi[all]
pyjwt
passlib[bcrypt]
sqlmodel
alembic 
psycopg2-binary
python-jose
pytest 
pytest-asyncio
httpx
pre-commit
black
flake8
